requirements.txt
# Core Dependencies - Compatible versions (Feb 2026)
langchain>=0.3.0,<1.0.0
langchain-core>=0.3.0,<1.0.0
langchain-community>=0.3.0,<1.0.0
langchain-text-splitters>=0.3.0,<1.0.0

# Default Provider: OpenAI (also works with OpenAI-compatible APIs like LiteLLM proxy)
langchain-openai>=0.3.0,<1.0.0

# 🔥 LiteLLM - Universal interface for 100+ providers (RECOMMENDED)
# Note: If using LiteLLM proxy server, you don't need to install this client-side
# litellm>=1.30.0

# Optional Providers (install as needed)
# Anthropic Claude
# langchain-anthropic>=0.1.0

# Google Gemini
# langchain-google-genai>=1.0.0

# HuggingFace Models
langchain-huggingface>=0.3.0,<1.0.0

# Cohere
# langchain-cohere>=0.1.0

# AWS Bedrock
# langchain-aws>=0.1.0

# HuggingFace Dataset
datasets>=4.0.0
huggingface-hub>=0.36.0
tokenizers>=0.21,<0.22  # Required by transformers

# Vector Store
faiss-cpu>=1.7.0
numpy>=1.24.0,<2.0.0

# Embeddings
sentence-transformers>=2.5.0  # Local embeddings
transformers>=4.0.0  # Required by sentence-transformers

# API Service
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1

# Security
slowapi==0.1.9  # Rate limiting
cryptography==42.0.2  # Encryption utilities

# Monitoring
prometheus-client==0.19.0

# Tools
duckduckgo-search>=4.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Development Dependencies (optional)
# Security Auditing (dev/CI)
# pip-audit>=2.7.0
# safety>=3.0.0

# Testing
# pytest>=7.4.0
# pytest-cov>=4.1.0

# Code Quality
# black>=23.12.0
# flake8>=7.0.0
# mypy>=1.8.0
